pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
include = '\.pyi?$'

[tool.mypy]
python_version = 3.11
ignore_missing_imports = true

[tool.poetry]
name = "rgfn"
version = "0.0.1"
description = "Scalable and Cost-Efficient de Novo Template-Based Molecular Generation"
authors = ["Piotr Gaiński", "Oussama Boussif", "Andrei Rekesh", "Dmytro Shevchuk", "Ali Parviz", "Mike Tyers", "Robert A. Batey", "Michał Koziarski"]

[tool.poetry.dependencies]
torch-geometric = '2.5.3'
torchmetrics = '1.2.0'
dgllife = '0.3.2'
tqdm = '4.67.1'
wandb = '0.15.12'
typing-extensions = '4.8.0'
pre-commit = '4.0.1'
omegaconf = '2.3.0'
gin-config = '0.5.0'
more-itertools = '10.5.0'
pandas = '2.1.4'
packaging = '24.1'
rdkit = '2023.9.5'
pydantic = '2.6.3'
pytest = '8.3.5'
jupyter = '1.0.0'
papermill = '2.4.0'
numpy = '1.26.4'
openpyxl = '3.1.4'
openbabel-wheel = '3.1.1.20'
meeko = '0.5.1'
wurlitzer = '3.1.0'
xlsxwriter = '3.2.0'
PyTDC = '1.1.15'
Biopython = '1.84.0'
